Move couple methods from BaseBasket to OrderSource
These could very well be in OrderSource, since there is no Front related
things in these.  Also, some functionality might need these, e.g.
basket campaings will need order_source.product_ids if trying to create
an order from an OrderSource with active basket campaing in database.

Noted this while implementing tests for discount line
taxing (SHOOP-2077).

Refs SHOOP-2077
